+++ src/main/kotlin/net/usbwire/usbplus/features/Compass.kt
@@ -1,11 +1,18 @@
 package net.usbwire.usbplus.features
 
-import ItemHelper
 import net.usbwire.usbplus.USBPlus
 import net.usbwire.usbplus.config.Config
 import net.usbwire.usbplus.util.Util
 import net.usbwire.usbplus.util.chat.Coordinates
+import net.usbwire.usbplus.commands.CompassCommand
+import gg.essential.api.EssentialAPI
 
+/**
+ * Monumenta compass helper! Gets where compass is pointing on left click
+ * Ported directly from U5B/jsmacros
+ * TODO: add command /compass to get position easily
+ * TODO: make a configurable keybind
+ */
 object Compass {
 	fun getCompass(): Coordinates.Coordinates {
 		val spawnPos = USBPlus.mc.world!!.spawnPos
@@ -22,8 +29,15 @@ object Compass {
 		Util.chat(message)
 	}
 
-	var clicked = false
+	fun configChanged(value: Boolean = Config.poiEnabled) {
+		if (value == true) {
+			CompassCommand.register()
+		} else if (value == false) {
+			EssentialAPI.getCommandRegistry().unregisterCommand(CompassCommand)
+		}
+	}
 
+	var clicked = false
 	fun onWorldTick() {
 		if (Config.compassEnabled == false) return
 		val click = USBPlus.mc.mouse.wasLeftButtonClicked()